[move] Node widget のフォルダ階層を変更しました.
